+++ lib/assets/components/toggleButton.dart
@@ -1,0 +1,11 @@
+import 'package:flutter/material.dart';
+
+class CustomToggleButton extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return ToggleButtons(
+      children: <Widget>[],
+      onPressed: (int index) {},
+    );
+  }
+}