+++ lib/services/auth.dart
@@ -14,6 +14,7 @@ import 'package:app/services/device_info_service.dart';
 class AuthError {
   static const unauthorized = "unauthorized";
   static const internal = "internal";
+  static const request = "request";
   static const unknown = "unknown";
   static const tokenExpired = "tokenExpired";
 }
@@ -27,11 +28,6 @@ class AuthenticationService {
   User _currentUser;
   User get currentUser => _currentUser;
 
-  bool logout() {
-    _currentUser = null;
-    return true;
-  }
-
   // This is not final as it can be changed
   String domainPrefix = "https://api.now-u.com/api/v1/";
 
@@ -42,14 +38,21 @@ class AuthenticationService {
   Future<bool> isUserLoggedIn() async {
     User user = await _sharedPreferencesService.loadUserFromPrefs();
     if (user != null) {
+      print("[isUserLoggedIn()]: user from Prefs not null");
       await _updateUser(user.getToken());
     }
+    print("[isUserLoggedIn()]: _currentUser: " + _currentUser.toString());
     return _currentUser != null;
   }
 
   Future _updateUser(String token) async {
     if (token != null) {
+      print("[_updateUser()]: _currentUser BEFORE getUser(token): " +
+          _currentUser.toString());
       _currentUser = await getUser(token);
+
+      print("[_updateUser()]: _currentUser AFTER getUser(token): " +
+          _currentUser.toString());
       _sharedPreferencesService.saveUserToPrefs(_currentUser);
     }
   }
@@ -74,11 +77,6 @@ class AuthenticationService {
   Future sendSignInWithEmailLink(
       String email, String name, bool acceptNewletter) async {
     try {
-      print("email | $email");
-      print("name | $name");
-      print("nl | $acceptNewletter");
-      print("operatingSystem | ${_deviceInfoService.osType}");
-      print("operatingSystemVersion | ${await _deviceInfoService.osVersion}");
       await http.post(
         domainPrefix + 'users',
         headers: <String, String>{
@@ -122,13 +120,20 @@ class AuthenticationService {
       }
 
       User user = await getUser(json.decode(response.body)['data']['token']);
-
-      await _updateUser(user.getToken());
+      _updateUser(user.getToken());
       return null;
     } on Error catch (e) {
       print("Error ${e.toString()}");
       return AuthError.unknown;
     }
+  }
+
+  bool logout() {
+    print("current user: " + _currentUser.toString());
+    _sharedPreferencesService.removeUserFromPrefs();
+    _currentUser = null;
+    print("current user: " + _currentUser.toString());
+    return true;
   }
 
   Future<User> getUser(String token) async {
@@ -143,20 +148,92 @@ class AuthenticationService {
     return u;
   }
 
-  Future<User> updateUserDetails(User user) async {
-    http.Response response = await http.put(
-      domainPrefix + 'users/me',
-      headers: <String, String>{
-        'Content-Type': 'application/json; charset=UTF-8',
-        'token': user.getToken(),
-      },
-      body: json.encode(user.getPostAttributes()),
-    );
-    if (handleAuthRequestErrors(response) != null) {
-      return handleAuthRequestErrors(response);
+  Future<bool> updateUserDetails({
+    String name,
+    DateTime dob,
+    String orgCode,
+  }) async {
+    try {
+      final Map<String, String> userDetials = {};
+      if (name != null) {
+        userDetials["full_name"] = name;
+      }
+      if (dob != null) {
+        userDetials["date_of_birth"] = dob.toString();
+      }
+      if (orgCode != null) {
+        userDetials["organisation_code"] = orgCode;
+      }
+      http.Response response = await http.put(
+        domainPrefix + 'users/me',
+        headers: <String, String>{
+          'Content-Type': 'application/json; charset=UTF-8',
+          'token': currentUser.getToken(),
+        },
+        body: jsonEncode(userDetials),
+      );
+      if (response.statusCode >= 300) {
+        return false;
+      }
+      print(response.body);
+      print(json.decode(response.body)['data']['organisation']);
+      User u = User.fromJson(json.decode(response.body)['data']);
+      currentUser.setName(u.getName());
+      currentUser.setDateOfBirth(u.getDateOfBirth());
+      print("new user og is ${u.organisation}");
+      currentUser.setOrganisation = u.organisation;
+      return true;
+    } catch (e) {
+      return false;
     }
-    User u = User.fromJson(json.decode(response.body));
-    return u;
+  }
+
+  Future<String> deleteUserAccount() async {
+    try {
+      http.Response response = await http.delete(
+        domainPrefix + 'users/me',
+        headers: <String, String>{
+          'Content-Type': 'application/json; charset=UTF-8',
+          'token': currentUser.getToken(),
+        },
+      );
+      if (response.statusCode >= 400 && response.statusCode < 500) {
+        return AuthError.request;
+      }
+      if (response.statusCode >= 500 && response.statusCode < 600) {
+        return AuthError.internal;
+      }
+      if (response.statusCode == 200) {
+        return null;
+      }
+      return AuthError.unknown;
+    } catch (e) {
+      return AuthError.unknown;
+    }
+  }
+
+  Future<bool> leaveOrganisation() async {
+    try {
+      print("Leaving org");
+      http.Response response = await http.put(
+        domainPrefix + 'users/me/organisations',
+        headers: <String, String>{
+          'Content-Type': 'application/json; charset=UTF-8',
+          'token': currentUser.getToken(),
+        },
+      );
+      if (response.statusCode >= 300) {
+        print("Put organisation_code sjdfkljsdf");
+        return false;
+      }
+      print("Put organisation_code null");
+      User u = User.fromJson(json.decode(response.body));
+      currentUser.setOrganisation = u.organisation;
+      return true;
+    } catch (e) {
+      print("Error");
+      return false;
+    }
   }
 
   Future<bool> completeAction(int actionId) async {